requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
requests
python-dotenv
pandas
numpy
ta
gspread
oauth2client
scikit-learn
